✏️ Rotation

+++ src/transform.c
@@ -6,7 +6,7 @@
 /*   By: sguzman <sguzman@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/01 08:28:33 by sguzman           #+#    #+#             */
-/*   Updated: 2024/03/03 13:48:56 by sguzman          ###   ########.fr       */
+/*   Updated: 2024/03/05 09:39:37 by sguzman          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,7 @@ void	zoom(t_scene *scene, int factor)
 	(*scene).refresh = 1;
 }
 
-void	rotate(t_scene *scene, int rx, int ry)
+void	rotate(t_scene *scene, float rx, float ry)
 {
 	(*scene).rotation.i += rx;
 	(*scene).rotation.j += ry;
